requirements.txt
altgraph @ file:///F:/SalvationLitePackage/whl/altgraph-0.16.1-py2.py3-none-any.whl
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
bootstrap4==0.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels @ file:///C:/Users/admin/Downloads/channels-4.0.0-py3-none-any.whl
chardet @ file:///C:/Users/admin/Downloads/chardet-5.2.0-py3-none-any.whl
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configobj==5.0.8
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
crypto==1.4.1
Django==3.2.16
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-filter==23.2
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.4.2
easyprocess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
eventlet @ file:///C:/Users/admin/Downloads/eventlet-0.34.2-py3-none-any.whl
exceptiongroup==1.1.1
flower @ file:///C:/Users/admin/Downloads/flower-2.0.1-py2.py3-none-any.whl
future==0.17.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
idna==3.4
jdcal==1.4.1
kombu==5.3.1
macholib @ file:///F:/SalvationLitePackage/whl/macholib-1.11-py2.py3-none-any.whl
mouseinfo==0.1.3
mss==9.0.1
mysqlclient @ file:///F:/Salvation/mysqlclient-2.2.0-cp39-cp39-win_amd64.whl
Naked==0.1.32
openpyxl==2.5.4
outcome==1.2.0
pefile==2018.8.8
pillow==10.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.38
pyautogui==0.9.52
PyAutoIt @ file:///F:/SalvationLitePackage/whl/PyAutoIt-0.6.5-py3-none-any.whl
pycparser==2.21
pycryptodome==3.18.0
pygetwindow==0.0.9
pygments==2.17.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2023.10
pymsgbox==1.0.9
PyMySQL==1.0.2
pyodbc @ file:///F:/SalvationLitePackage/whl/pyodbc-4.0.35-cp39-cp39-win_amd64.whl
pyperclip==1.8.2
pyrect==0.2.0
pyscreenshot==3.1
pyscreeze==0.1.29
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
pywin32-ctypes @ file:///F:/SalvationLitePackage/whl/pywin32_ctypes-0.2.0-py2.py3-none-any.whl
PyYAML==6.0
redis @ file:///C:/Users/admin/Downloads/redis-5.0.1-py3-none-any.whl
requests==2.31.0
robotframework @ file:///F:/SalvationLitePackage/whl/robotframework-3.1.1-py2.py3-none-any.whl
robotframework-selenium2library @ file:///F:/SalvationLitePackage/whl/robotframework_selenium2library-3.0.0-py2.py3-none-any.whl
robotframework-selenium2screenshots @ file:///F:/SalvationLitePackage/whl/robotframework_selenium2screenshots-0.8.1-py2.py3-none-any.whl
robotframework-seleniumlibrary @ file:///F:/SalvationLitePackage/whl/robotframework_seleniumlibrary-3.3.1-py2.py3-none-any.whl
selenium @ file:///F:/SalvationLitePackage/whl/selenium-3.12.0-py2.py3-none-any.whl
shellescape==3.8.1
simpleui==4.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.4
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
uvicorn @ file:///C:/Users/admin/Downloads/uvicorn-0.25.0-py3-none-any.whl
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
xlrd @ file:///D:/app/xlrd-1.1.0-py2.py3-none-any.whl
